data_extraction_module/requirements.txt
beautifulsoup4==4.2.1
lxml==3.2.3
python-dwca-reader==0.6.1
